+++ packages/account-tree-controller/CHANGELOG.md
@@ -7,8 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `group.type` tag ([#6214](https://github.com/MetaMask/core/pull/6214))
+  - This `type` can be used as a tag to strongly-type (tagged-union) the `AccountGroupObject`.
+- Add `group.metadata` metadata object ([#6214](https://github.com/MetaMask/core/pull/6214))
+  - Given the `group.type` you will now have access to specific metadata information (e.g. `groupIndex` for multichain account groups)
+
 ### Changed
 
+- **BREAKING:** Bump peer dependency `@metamask/account-api` from `^0.3.0` to `^0.7.0` ([#6214](https://github.com/MetaMask/core/pull/6214)), ([#6216](https://github.com/MetaMask/core/pull/6216)), ([#6222](https://github.com/MetaMask/core/pull/6222))
+- **BREAKING:** Move `wallet.metadata.type` tag to `wallet` node ([#6214](https://github.com/MetaMask/core/pull/6214))
+  - This `type` can be used as a tag to strongly-type (tagged-union) the `AccountWalletObject`.
 - Defaults to the EVM account from a group when using `setSelectedAccountGroup` ([#6208](https://github.com/MetaMask/core/pull/6208))
   - In case no EVM accounts are found in a group (which should not be possible), it will defaults to the first account of that group.